+++ src/chains/evm/hub/constants/chain.ts
@@ -80,7 +80,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.USDC,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.USDC],
         poolAddress: convertToGenericAddress("0x88f15e36308ED060d8543DA8E2a5dA0810Efded2" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.STABLECOIN_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.AVAX]: {
         token: {
@@ -160,7 +164,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.wBTC_eth,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.wBTC_eth],
         poolAddress: convertToGenericAddress("0x9936812835476504D6Cf495F4F0C718Ec19B3Aff" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.BTCb_ava]: {
         token: {
@@ -170,7 +178,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.BTCb_ava,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.BTCb_ava],
         poolAddress: convertToGenericAddress("0x1C51AA1516e1156d98075F2F64e259906051ABa9" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.cbBTC_base]: {
         token: {
@@ -180,7 +192,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.cbBTC_base,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.cbBTC_base],
         poolAddress: convertToGenericAddress("0x9eD81F0b5b0E9b6dE00F374fFc7f270902576EF7" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.BNB]: {
         token: {
@@ -214,7 +230,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.BTCB_bsc,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.BTCB_bsc],
         poolAddress: convertToGenericAddress("0xC2FD40D9Ec4Ae7e71068652209EB75258809e131" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.ETH_arb]: {
         token: {
@@ -250,7 +270,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.SolvBTC,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.SolvBTC],
         poolAddress: convertToGenericAddress("0x307bCEC89624660Ed06C97033EDb7eF49Ab0EB2D" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.JOE]: {
         token: {
@@ -296,7 +320,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.POL,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.POL],
         poolAddress: convertToGenericAddress("0x481cF0c02BF17a33753CE32f1931ED9990fFB40E" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.POL_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.wBTC_pol]: {
         token: {
@@ -306,7 +334,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.wBTC_pol,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.wBTC_pol],
         poolAddress: convertToGenericAddress("0x7054254933279d93D97309745AfbFF9310cdb570" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.wETH_pol]: {
         token: {
@@ -322,15 +354,19 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
           MAINNET_LOAN_TYPE_ID.ETH_EFFICIENCY,
         ]),
       },
-      [MAINNET_FOLKS_TOKEN_ID.aUSD]: {
+      [MAINNET_FOLKS_TOKEN_ID.aUSD_ava]: {
         token: {
           type: TokenType.ERC20,
           decimals: 6,
         },
-        folksTokenId: MAINNET_FOLKS_TOKEN_ID.aUSD,
-        poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.aUSD],
+        folksTokenId: MAINNET_FOLKS_TOKEN_ID.aUSD_ava,
+        poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.aUSD_ava],
         poolAddress: convertToGenericAddress("0xc7DdB440666c144c2F27a3a5156D636Bacfc769C" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.STABLECOIN_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.savUSD]: {
         token: {
@@ -340,7 +376,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.savUSD,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.savUSD],
         poolAddress: convertToGenericAddress("0xE6B7713854620076B5716E2743262D315bf8609D" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.STABLECOIN_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.wBTC_arb]: {
         token: {
@@ -350,7 +390,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.wBTC_arb,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.wBTC_arb],
         poolAddress: convertToGenericAddress("0x3445055F633fEF5A64F852aaCD6dA76143aCA109" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.tBTC_arb]: {
         token: {
@@ -360,7 +404,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.tBTC_arb,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.tBTC_arb],
         poolAddress: convertToGenericAddress("0xdd9eFBf83572f5387381aD3A04b1318221d545A2" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.BTC_EFFICIENCY,
+        ]),
       },
       [MAINNET_FOLKS_TOKEN_ID.wstETH_arb]: {
         token: {
@@ -437,7 +485,11 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: MAINNET_FOLKS_TOKEN_ID.MaticX,
         poolId: MAINNET_POOLS[MAINNET_FOLKS_TOKEN_ID.MaticX],
         poolAddress: convertToGenericAddress("0x59023eFDB22B9d8b2C7aeD842aC1fd2f6110e5B5" as EvmAddress, ChainType.EVM),
-        supportedLoanTypes: new Set([MAINNET_LOAN_TYPE_ID.DEPOSIT, MAINNET_LOAN_TYPE_ID.GENERAL]),
+        supportedLoanTypes: new Set([
+          MAINNET_LOAN_TYPE_ID.DEPOSIT,
+          MAINNET_LOAN_TYPE_ID.GENERAL,
+          MAINNET_LOAN_TYPE_ID.POL_EFFICIENCY,
+        ]),
       },
     } satisfies Record<MainnetFolksTokenId, HubTokenData>,
     rewards: {
@@ -491,6 +543,14 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
             token: {
               type: TokenType.ERC20,
               decimals: 6,
+            },
+          },
+          [MAINNET_REWARDS_TOKEN_ID.POL]: {
+            rewardTokenId: MAINNET_REWARDS_TOKEN_ID.POL,
+            nodeId: "0x5ffc96dd95bbfe846c600fdd21d33f36e7236fee9344bb8b3f4edada887d26d4" as NodeId,
+            token: {
+              type: TokenType.NATIVE,
+              decimals: 18,
             },
           },
         } satisfies Record<MainnetRewardsTokenId, HubRewardTokenData>,
@@ -639,6 +699,46 @@ export const HUB_CHAIN: Record<NetworkType, HubChain> = {
         folksTokenId: TESTNET_FOLKS_TOKEN_ID.MON,
         poolId: TESTNET_POOLS[TESTNET_FOLKS_TOKEN_ID.MON],
         poolAddress: convertToGenericAddress("0x48808cE8cC5c252d095E9D7b9c6c12FE181d835F" as EvmAddress, ChainType.EVM),
+        supportedLoanTypes: new Set([TESTNET_LOAN_TYPE_ID.DEPOSIT, TESTNET_LOAN_TYPE_ID.GENERAL]),
+      },
+      [TESTNET_FOLKS_TOKEN_ID.sMON]: {
+        token: {
+          type: TokenType.ERC20,
+          decimals: 18,
+        },
+        folksTokenId: TESTNET_FOLKS_TOKEN_ID.sMON,
+        poolId: TESTNET_POOLS[TESTNET_FOLKS_TOKEN_ID.sMON],
+        poolAddress: convertToGenericAddress("0xbaEe7dD681BF1652171ebEE4c96da78407922209" as EvmAddress, ChainType.EVM),
+        supportedLoanTypes: new Set([TESTNET_LOAN_TYPE_ID.DEPOSIT, TESTNET_LOAN_TYPE_ID.GENERAL]),
+      },
+      [TESTNET_FOLKS_TOKEN_ID.aprMON]: {
+        token: {
+          type: TokenType.ERC20,
+          decimals: 18,
+        },
+        folksTokenId: TESTNET_FOLKS_TOKEN_ID.aprMON,
+        poolId: TESTNET_POOLS[TESTNET_FOLKS_TOKEN_ID.aprMON],
+        poolAddress: convertToGenericAddress("0x6B5a247C436DfA5b5e0b22C83EFc709672C7B337" as EvmAddress, ChainType.EVM),
+        supportedLoanTypes: new Set([TESTNET_LOAN_TYPE_ID.DEPOSIT, TESTNET_LOAN_TYPE_ID.GENERAL]),
+      },
+      [TESTNET_FOLKS_TOKEN_ID.gMON]: {
+        token: {
+          type: TokenType.ERC20,
+          decimals: 18,
+        },
+        folksTokenId: TESTNET_FOLKS_TOKEN_ID.gMON,
+        poolId: TESTNET_POOLS[TESTNET_FOLKS_TOKEN_ID.gMON],
+        poolAddress: convertToGenericAddress("0x503EDEc647AfddDC8E1D7DE13780db9B1388a4Fc" as EvmAddress, ChainType.EVM),
+        supportedLoanTypes: new Set([TESTNET_LOAN_TYPE_ID.DEPOSIT, TESTNET_LOAN_TYPE_ID.GENERAL]),
+      },
+      [TESTNET_FOLKS_TOKEN_ID.shMON]: {
+        token: {
+          type: TokenType.ERC20,
+          decimals: 18,
+        },
+        folksTokenId: TESTNET_FOLKS_TOKEN_ID.shMON,
+        poolId: TESTNET_POOLS[TESTNET_FOLKS_TOKEN_ID.shMON],
+        poolAddress: convertToGenericAddress("0xF494F0Ba4e447a5FCaEacB459Cbd8cBD2F94a9Ac" as EvmAddress, ChainType.EVM),
         supportedLoanTypes: new Set([TESTNET_LOAN_TYPE_ID.DEPOSIT, TESTNET_LOAN_TYPE_ID.GENERAL]),
       },
     } satisfies Record<TestnetFolksTokenId, HubTokenData>,